requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
faiss-cpu
beautifulsoup4
requests
python-dotenv
tqdm
